requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.23.2
mysql-connector-python==8.2.0
redis==5.0.1
websockets==12.0
kafka-python==2.0.2
pydantic==2.4.2

# Frontend Dependencies
streamlit==1.28.1
pandas==2.1.2
folium==0.14.0
streamlit-folium==0.15.0

# External Integrations

# Development Tools
python-dotenv==1.0.0
pathlib==1.0.1